+++ src/variables/helpMessages.js
@@ -43,13 +43,15 @@ const HELP = {
     args: {},
   },
   copy: {
-    usage: "copy <src> <dest> [-t] [--rm-symlink]",
+    usage: "copy <src> <dest> [-y] [-t] [--rm-symlink]",
     desc: "Copies a file/directory from the source to the destination (with its contents). Absolute and relative paths for both arguments are accepted. The destination must have the name of the copied file/directory (otherwise if the path already exists, it will be overwritten).",
     args: {
       "<src>":
         "The source file/directory. The file/directory must exist, otherwise it will throw an error.",
       "<dest>":
         "The destination file/directory with its (new) name. If the name passed already exists, it will be overwritten.",
+      "[-y]":
+        "If the destination exists, then automatically accept the prompt instead of confirming to overwrite the file/directory.",
       "[-t]":
         "Only for copying directories. Keeps the original timestamps of the nested files/directories. By default, this is false.",
       "[--rm-symlink]":